Add files via upload
We have reduced a single use case. The company technician can no longer check the status of the weather tower. Our implementation felt very simple, and we did not really instantiate a WeatherTechnician or anything to actually check and use the method, so there was no point to having it.