Add files via upload
Lots of small changes to match actual implementation details, but the core of the program is the same. Adjusted some of the interface implementations, like removing TowerCommunication's "StatusReporter" because we removed the use case for technicians to check the status. Little things like which fields were called what, and how we constructed the objects so they had references to each other as well.